+++ scripts/ard/publishers.js
@@ -10,7 +10,7 @@
 const undici = require('../../packages/undici')()
 
 // init storage
-const Storage = require('../../packages/storage-wrapper')
+const Storage = require('../../packages/storage')
 
 const storage = new Storage({
 	gs: { projectId: process.env.GCP_PROJECT_ID },